Step 1: moved code for visualization of ephemeris into separate class
