+++ src/gui/AstroCalcExtraEphemerisDialog.cpp
@@ -53,12 +53,12 @@ void AstroCalcExtraEphemerisDialog::createDialogContent()
 
 	connect(ui->skipDataCheckBox, SIGNAL(clicked()), this, SLOT(setOptionStatus()));
 
-	connectBoolProperty(ui->skipDataCheckBox,	"SolarSystem.ephemerisSkippedData");
-	connectBoolProperty(ui->skipMarkersCheckBox,	"SolarSystem.ephemerisSkippedMarkers");
-	connectIntProperty(ui->dataStepSpinBox,		"SolarSystem.ephemerisDataStep");
-	connectBoolProperty(ui->smartDatesCheckBox,	"SolarSystem.ephemerisSmartDates");
-	connectBoolProperty(ui->scaleMarkersCheckBox,	"SolarSystem.ephemerisScaleMarkersDisplayed");
-	connectIntProperty(ui->lineThicknessSpinBox,	"SolarSystem.ephemerisLineThickness");
+	connectBoolProperty(ui->skipDataCheckBox,	"EphemerisMgr.ephemerisSkippedData");
+	connectBoolProperty(ui->skipMarkersCheckBox,	"EphemerisMgr.ephemerisSkippedMarkers");
+	connectIntProperty(ui->dataStepSpinBox,		"EphemerisMgr.ephemerisDataStep");
+	connectBoolProperty(ui->smartDatesCheckBox,	"EphemerisMgr.ephemerisSmartDates");
+	connectBoolProperty(ui->scaleMarkersCheckBox,	"EphemerisMgr.ephemerisScaleMarkersDisplayed");
+	connectIntProperty(ui->lineThicknessSpinBox,	"EphemerisMgr.ephemerisLineThickness");
 
 	setOptionStatus();
 }
